requirements.txt
vimeo_downloader
yt-dlp
openai-whisper